fix:move file position and support partition table
