+++ pkg/runtime/optimize/dal/analyze_table.go
@@ -15,10 +15,11 @@
  * limitations under the License.
  */
 
-package ddl
+package dal
 
 import (
 	"context"
+	"github.com/arana-db/arana/pkg/runtime/plan/dal"
 )
 
 import (
@@ -26,7 +27,6 @@ import (
 	"github.com/arana-db/arana/pkg/proto/rule"
 	"github.com/arana-db/arana/pkg/runtime/ast"
 	"github.com/arana-db/arana/pkg/runtime/optimize"
-	"github.com/arana-db/arana/pkg/runtime/plan/ddl"
 )
 
 func init() {
@@ -41,5 +41,5 @@ func optimizeAnalyzeTable(ctx context.Context, o *optimize.Optimizer) (proto.Pla
 	}
 
 	stmt := o.Stmt.(*ast.AnalyzeTableStatement)
-	return ddl.NewAnalyzeTablePlan(stmt, shards), nil
+	return dal.NewAnalyzeTablePlan(stmt, shards), nil
 }
